+++ CHANGELOG.rst
@@ -15,6 +15,7 @@ Added
     - Setting ``PROJECTROLES_ENABLE_SEARCH`` (#70)
     - Re-enable "home" link in project breadcrumb (#80)
     - ``get_extra_data_link()`` in ProjectAppPluginPoint for timeline extra data (#6)
+    - Allow overriding project class in ProjectAccessMixin
 - **Taskflowbackend**
     - ``test_mode`` flag configured with ``TASKFLOW_TEST_MODE`` in settings (#67)
 - **Timeline**